add only owner can delete/update

+++ app/routes/post.py
@@ -36,8 +36,7 @@ def create_post(post : schemas.PostBase ,
     
 #    new_post = models.Post(title = post.title, content = post.content, published = post.published, )
 #    **post.dict() 
-   new_post = models.Post(**post.dict() )
-
+   new_post = models.Post( owner_id = get_current_user.id,**post.dict() )
    db.add(new_post)
    db.commit()
    db.refresh(new_post)
@@ -45,7 +44,9 @@ def create_post(post : schemas.PostBase ,
 
 
 @router.put("/{id}", response_model=schemas.PostResponse)
-def update_post(id:int, post: schemas.PostBase, db: Session = Depends(get_db)):
+def update_post(id:int, post: schemas.PostBase, 
+                db: Session = Depends(get_db),
+                get_current_user: int =Depends(oauth2.get_current_user)):
     
     current_post = db.query(models.Post).filter(models.Post.id == id)
     
@@ -54,6 +55,8 @@ def update_post(id:int, post: schemas.PostBase, db: Session = Depends(get_db)):
     if c_post == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"{id} not found")
     
+    if c_post.owner_id != get_current_user.id :
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"{id} post not belongs to you ")
     
     # post.update({'title': 'updated title', 'content': 'updated content'}, synchronize_session = False)
     current_post.update(post.dict(), synchronize_session = False)
@@ -74,6 +77,9 @@ def delete_post(id: int,
     if post == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"{id} is wrong id ")
     
+    if post.owner_id != get_current_user.id :
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"{id} post not belongs to you ")
+
     db.delete(post)
     db.commit()
